requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
idna==3.3
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
